pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ComfyUI-RMBG"
version = "1.1.0"
description = "A ComfyUI node for removing image backgrounds using RMBG-2.0"
authors = [{ name = "AILab" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0,<3.0.0",
    "torchvision>=0.15.0,<1.0.0",
    "Pillow>=9.0.0,<10.0.0",
    "numpy>=1.22.0,<2.0.0",
    "transformers>=4.30.0,<5.0.0",
    "safetensors>=0.3.0,<1.0.0",
    "timm>=0.6.12,<1.0.0",
    "huggingface-hub>=0.16.0,<1.0.0"
]

[tool.setuptools.packages.find]
include = ["ComfyUI-RMBG*"] 